Int is 32bit on non-64bit platforms (#564)

+++ Blockchain/Stellar/Models/StellarAccount.swift
@@ -18,9 +18,9 @@ struct StellarAccount {
     // Sequence is used when submitting a transaction
     // from the user's account.
     // [Read more here:](https://www.stellar.org/developers/guides/concepts/fees.html#minimum-account-balance " Minimum Account Balance")
-    let sequence: Int
+    let sequence: Int64
     
-    let subentryCount: Int
+    let subentryCount: UInt
 }
 
 extension StellarAccount {
